Exclusive no longer implies auto-delete on queue.declare.
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid@583821 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qpid/broker/SemanticState.cpp
@@ -290,7 +290,7 @@ void SemanticState::ConsumerImpl::cancel()
 {
     if(queue) {
         queue->cancel(this);
-        if (queue->canAutoDelete()) {            
+        if (queue->canAutoDelete() && !queue->hasExclusiveOwner()) {            
             parent->getSession().getBroker().getQueues().destroyIf(
                 queue->getName(), 
                 boost::bind(boost::mem_fn(&Queue::canAutoDelete), queue));
